Update DatabaseManager.java
Fixed an error where "true" was being passed as a parameter in the insert method. This is because I removed a parameter from the method, but the last parameter is of the type Object... so it can accept any number of inputs, so when I removed the parameter before it, the value got caught in the Object... param.